+++ Assets/Adjust/Unity/AdjustUtils.cs
@@ -41,6 +41,7 @@ namespace com.adjust.sdk
         public static string KeyTestOptionsTeardown = "teardown";
         public static string KeyTestOptionsNoBackoffWait = "noBackoffWait";
         public static string KeyTestOptionsiAdFrameworkEnabled = "iAdFrameworkEnabled";
+        public static string KeyTestOptionsAdServicesFrameworkEnabled = "adServicesFrameworkEnabled";
 
         public static int ConvertLogLevel(AdjustLogLevel? logLevel)
         {