+++ Assets/Adjust/iOS/Adjust.h
@@ -2,7 +2,7 @@
 //  Adjust.h
 //  Adjust
 //
-//  V4.24.0
+//  V4.26.1
 //  Created by Christian Wellenbrock (wellle) on 23rd July 2013.
 //  Copyright © 2012-2017 Adjust GmbH. All rights reserved.
 //
@@ -11,6 +11,7 @@
 #import "ADJConfig.h"
 #import "ADJAttribution.h"
 #import "ADJSubscription.h"
+#import "ADJThirdPartySharing.h"
 
 @interface AdjustTestOptions : NSObject
 
@@ -26,6 +27,7 @@
 @property (nonatomic, assign) BOOL deleteState;
 @property (nonatomic, assign) BOOL noBackoffWait;
 @property (nonatomic, assign) BOOL iAdFrameworkEnabled;
+@property (nonatomic, assign) BOOL adServicesFrameworkEnabled;
 @property (nonatomic, assign) BOOL enableSigning;
 @property (nonatomic, assign) BOOL disableSigning;
 
@@ -43,6 +45,7 @@ extern NSString * __nonnull const ADJEnvironmentProduction;
 extern NSString * __nonnull const ADJAdRevenueSourceMopub;
 extern NSString * __nonnull const ADJAdRevenueSourceAdmob;
 extern NSString * __nonnull const ADJAdRevenueSourceFbNativeAd;
+extern NSString * __nonnull const ADJAdRevenueSourceFbAudienceNetwork;
 extern NSString * __nonnull const ADJAdRevenueSourceIronsource;
 extern NSString * __nonnull const ADJAdRevenueSourceFyber;
 extern NSString * __nonnull const ADJAdRevenueSourceAerserv;
@@ -273,6 +276,10 @@ extern NSString * __nonnull const ADJUrlStrategyChina;
  */
 + (void)disableThirdPartySharing;
 
++ (void)trackThirdPartySharing:(nonnull ADJThirdPartySharing *)thirdPartySharing;
+
++ (void)trackMeasurementConsent:(BOOL)enabled;
+
 /**
  * @brief Track subscription.
  *
@@ -281,6 +288,10 @@ extern NSString * __nonnull const ADJUrlStrategyChina;
 + (void)trackSubscription:(nonnull ADJSubscription *)subscription;
 
 + (void)requestTrackingAuthorizationWithCompletionHandler:(void (^_Nullable)(NSUInteger status))completion;
+
++ (int)appTrackingAuthorizationStatus;
+
++ (void)updateConversionValue:(NSInteger)conversionValue;
 
 /**
  * Obtain singleton Adjust object.
@@ -342,5 +353,9 @@ extern NSString * __nonnull const ADJUrlStrategyChina;
 - (nullable NSURL *)convertUniversalLink:(nonnull NSURL *)url scheme:(nonnull NSString *)scheme;
 
 - (void)requestTrackingAuthorizationWithCompletionHandler:(void (^_Nullable)(NSUInteger status))completion;
+
+- (int)appTrackingAuthorizationStatus;
+
+- (void)updateConversionValue:(NSInteger)conversionValue;
 
 @end